requirements.txt
alabaster==0.7.10
appdirs==1.4.3
asn1crypto==0.24.0
Babel==2.6.0
bcrypt==3.1.4
certifi==2017.7.27.1
cffi==1.11.4
cfscrape==1.9.1
chardet==3.0.4
colorclass==2.2.0
cryptography==2.1.4
Django==2.0.6
django-cors-headers==2.1.0
djangorestframework==3.7.1
docopt==0.6.2
docutils==0.14
formats==0.1.1
future==0.16.0
idna==2.6
imagesize==1.0.0
Jinja2==2.10
MarkupSafe==1.0
mouse==0.7.0
#mysqlclient==1.3.12
netaddr==0.7.19
packaging==16.8
paramiko==2.4.0
passlib==1.7.1
pbr==3.1.1
pyasn1==0.4.2
pycparser==2.18
PyExecJS==1.5.0
Pygments==2.2.0
PyMySQL==0.8.0
PyNaCl==1.2.1
pyparsing==2.2.0
python3-xlib==0.15
pytz==2018.3
requests==2.18.4
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.1
sphinxcontrib-websupport==1.0.1
tortilla==0.4.2
urllib3==1.22
uWSGI==2.0.17
